chore: remove unused files relasted to deposit form
